requirements.txt
numpy>=1.16.4
cffi==1.11.5

matplotlib
tensorflow
gym
tqdm

stable-baselines3
box2d-py==2.3.8
pybullet
gym-minigrid	
scikit-optimize
optuna
pytablewriter
seaborn
pyyaml>=5.1
# sb3-contrib>=1.0

scipy>=0.18.0
numba>=0.49.1
numpy-quaternion